backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
supabase==2.3.1
aiohttp==3.9.1
feedparser==6.0.10
web3==6.11.0
python-multipart==0.0.6
schedule==1.2.0
pydantic==2.5.0
lxml==4.9.3
langdetect==1.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-generativeai>=0.3.0
resend>=0.7.0